+++ scripts/deskhpsdr-install-deps-on-current-image.sh
@@ -95,6 +95,7 @@ DEBIAN_PACKAGES=(
   git
   cmake
   autoconf
+  perl
   autopoint
   gettext
   automake